recommender/requirements.txt
django == 2.1.2
requests == 2.20.0
djangorestframework
